requirements.txt
discord.py>=2.3,<3.0
anthropic>=0.39,<1.0
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
aiosqlite>=0.20,<1.0
python-dotenv>=1.0,<2.0
